Split main window controller
